+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Version History
 
-- 0.1.3: Added [pyawsopstoolkit.models](pyawsopstoolkit/models) package. (latest)
+- 0.1.4: Added [pyawsopstoolkit.advsearch](pyawsopstoolkit/advsearch) package. (latest)
+- 0.1.3: Added [pyawsopstoolkit.models](pyawsopstoolkit/models) package.
 - 0.1.2: Added Account, Credentials, and Session.
 - 0.1.1: Added [pyawsopstoolkit.validators](pyawsopstoolkit/validators) package.
 - 0.1.0: Added [pyawsopstoolkit.exceptions](pyawsopstoolkit/exceptions) package.